requirements.txt
Django==1.10.1
django-filter==0.15.2
djangorestframework==3.4.7
Markdown==2.6.7
MySQL-python==1.2.5
Pillow==2.9.0
wheel==0.24.0
